Trivial sklearn classifiers.
