+++ Code/classifiers/svm.py
@@ -1,14 +1,16 @@
-from Code.classifiers.classifier import Classifier
+from sklearn import svm
+from Code.classifiers.SKClassifier import SKClassifier
 
 
-class SVMClassifier(Classifier):
-
-    def __init__(self, c: float):
+class SVMClassifier(SKClassifier):
+    def __init__(self, penalty):
         super().__init__()
-        self.c = c
+        self.model = svm.SVC(kernel="linear", C=penalty)
 
-    def train(self, train_data):
-        return super().train(train_data)
 
-    def test(self, test_data):
-        return super().test(test_data)
+if __name__ == '__main__':
+    for penalty in [1/2, 1, 3/2, 2]:
+        svmc = SVMClassifier(penalty=penalty)
+        train_set = svmc.read_data_from_file("DataSet/phase2/phase2_train.csv")
+        svmc.train(train_set)
+        svmc.test(svmc.read_data_from_file("DataSet/phase2/phase2_test.csv"))
